* Ensured data gets added to session for depth, vessel name/construction/year, salvage services * Rounded down location radius number * Vessel depth now an int

+++ api/routes/location.js
@@ -412,7 +412,7 @@ export default function (app) {
       session['location-standard'].longitude = parseFloat(
         session['location-standard'].longitude
       );
-      session['location-standard'].radius = parseFloat(
+      session['location-standard'].radius = parseInt(
         session['location-standard'].radius
       );
